+++ src/test/java/blackjack/domain/DeckTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
 
 public class DeckTest {
